+++ presentation/cli/output.go
@@ -1,6 +1,19 @@
 package cli
 
-import "time"
+import (
+	"fmt"
+	"time"
+)
+
+func formatDuration(d time.Duration) string {
+	if d < time.Minute {
+		return fmt.Sprintf("%ds", int(d.Seconds()))
+	}
+	if d < time.Hour {
+		return fmt.Sprintf("%dm%ds", int(d.Minutes()), int(d.Seconds())%60)
+	}
+	return fmt.Sprintf("%dh%dm", int(d.Hours()), int(d.Minutes())%60)
+}
 
 // event is the JSON shape of an event in CLI output.
 //